fix(pipeline): PipelineIterator.get_state returns the position as an int
`get_state` documents a JSON-serializable dict, and `set_state` reads the position with
`int(...)`, but the position came back as `np.int64`: `json.dumps(iterator.get_state())` raised
`TypeError: Object of type int64 is not JSON serializable`, and a checkpoint store that builds
its restore template from the state (Orbax's `construct_restore_args`) refused the NumPy scalar.
The commit that introduced the state named JSON serializability as its purpose and gave no
reason for the NumPy type; the test that pinned `np.int64` asserted the implementation against
the documented contract.

The position is a Python `int`. The test asserts `int`, and a new test checks that the state
resumes the same batches after a `json.dumps`/`json.loads` round trip. The comparison quick
reference's printed checkpoint reads `'position': 8` instead of `'position': np.int64(8)` in
the script, its notebook and its docs page.

+++ src/datarax/pipeline/iteration.py
@@ -37,7 +37,6 @@ from typing import Any, TYPE_CHECKING
 
 import jax
 import jax.numpy as jnp
-import numpy as np
 from flax import nnx
 
 from datarax.core.operator import OperatorModule
@@ -439,7 +438,7 @@ class PipelineIterator:
         """
         counts = [int(self._live_variables[index].get_value()) for index in self._rng_count_indices]
         return {
-            "position": np.int64(self._position),
+            "position": int(self._position),
             "epoch": int(self._live_variables[self._epoch_index].get_value()),
             "rng_counts": counts,
             "version": _ITERATOR_STATE_VERSION,
